requirements.txt
aiogram
aiosqlite
python-dotenv
pandas
